restore old ZGC logging
